refactor: 주석 제거 및 attributePath변경
